dsl/difyforsiliconflow/bizyair/requirements.txt
uvicorn== 0.34.0
fastapi== 0.115.6
cos-python-sdk-v5==1.9.33
python-multipart
pydantic
ipython==8.31.0
websocket-client==1.8.0